Single Execution/requirements.txt
python-docx==0.8.11
pandas==1.4.2
requests==2.26.0
urllib3==1.26.7
chardet==4.0.0
charset_normalizer==2.0.6
google-auth==2.17.0
google-auth-oauthlib==1.0.0
openpyxl==3.1.2

